refactor: Co-locate feature-specific components
This refactoring improves the project's architecture by moving feature-specific components, utilities, and tests from global directories (`src/components`, `src/utils`) into their respective feature directories under `src/app`.

The primary changes include:
- Moved all components and utilities related to the "Registry" feature into `src/app/registry/components` and `src/app/registry/lib`.
- Moved all components related to the "Things to Do" feature into `src/app/things-to-do/components`.
- Moved the `Modal` component into the `registry` feature directory as it was only used there.
- Updated all import paths across the application to reflect the new file structure.

This co-location of files improves modularity, cohesion, and maintainability, making the codebase easier to navigate and scale.
