+++ src/app/registry/components/CategoryFilter.tsx
