+++ src/app/registry/components/PriceRangeFilter.tsx
