+++ src/app/registry/components/RegistryCardSkeleton.tsx
