+++ src/app/registry/components/RegistryItemProgressBar.tsx
