+++ src/app/registry/components/__tests__/CategoryFilter.test.tsx
