+++ src/app/registry/components/__tests__/PriceRangeFilter.test.tsx
