+++ src/app/registry/components/__tests__/RegistryCard.test.tsx
