+++ src/app/registry/components/__tests__/RegistryItemForm.test.tsx
