+++ src/app/registry/components/__tests__/RegistryItemProgressBar.test.tsx
