+++ src/app/registry/lib/__tests__/registryStatusUtils.test.ts
