+++ src/app/registry/lib/registryStatusUtils.ts
